# wip: prototype of "compiled factories"