--- a/pkg/sql/opt/exec/BUILD.bazel
+++ b/pkg/sql/opt/exec/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
     deps = [
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/descpb",  # keep
+        "//pkg/sql/idxusage",
         "//pkg/sql/inverted",
         "//pkg/sql/opt",  # keep
         "//pkg/sql/opt/cat",
@@ -21,6 +22,7 @@ go_library(
         "//pkg/sql/types",
         "//pkg/util/intsets",
         "//pkg/util/optional",
+        "@com_github_cockroachdb_errors//:errors",
     ],
 )