--- a/pkg/sql/opt/exec/explain/emit.go
+++ b/pkg/sql/opt/exec/explain/emit.go
@@ -293,6 +293,9 @@ func (e *emitter) nodeName(n *Node) (name string, _ error) {
 		}
 		return "scan", nil
 
+	case placeholderScanOp:
+		return "placeholder-scan", nil
+
 	case valuesOp:
 		a := n.args.(*valuesArgs)
 		switch {
@@ -706,6 +709,10 @@ func (e *emitter) emitNodeAttributes(ctx context.Context, evalCtx *eval.Context,
 		}
 		e.emitLockingPolicy(a.Params.Locking)
 		e.emitPolicies(ob, a.Table, n)
+
+	case placeholderScanOp:
+		// a := n.args.(*placeholderScanArgs)
+		e.emitPolicies(ob, nil, n)
 
 	case valuesOp:
 		a := n.args.(*valuesArgs)