This prototype memoizes repeated work when converting an optimized memo into a
physical plan. Placeholder scans now have their own constructors in exec
factories, allowing constraint building, which depends on placeholder values, to
move from the execbuild phase to the constructor phase. As a result, the
execbuild phase does not need to be repeated for each execution of placeholder
scan.

Data structures that do not depend on placeholder values can be reused across
executions of the same placeholder scan. These objects are cached in a closure
which is invoked on future executions to create an exec plan without creating or
invoking execbuilder or the factories directly. I'm calling these closures
"compiled factories" (naming is not yet consistent in this prototype).

The end-to-end benchmarks show the potential speedup and reduction in
allocations for a simple kv read.

```
name                                     old time/op    new time/op    delta
EndToEnd/kv-read/vectorize=on/Simple        164µs ± 2%     163µs ± 2%    ~     (p=0.565 n=20+20)
EndToEnd/kv-read/vectorize=on/Prepared      114µs ± 5%     109µs ± 3%  -4.59%  (p=0.000 n=20+20)
EndToEnd/kv-read/vectorize=off/Simple       178µs ± 4%     178µs ± 2%    ~     (p=0.892 n=20+17)
EndToEnd/kv-read/vectorize=off/Prepared     114µs ± 5%     107µs ± 6%  -5.68%  (p=0.000 n=20+19)

name                                     old alloc/op   new alloc/op   delta
EndToEnd/kv-read/vectorize=on/Simple       30.4kB ± 4%    31.6kB ± 5%  +3.87%  (p=0.000 n=17+19)
EndToEnd/kv-read/vectorize=on/Prepared     19.4kB ± 4%    18.6kB ± 5%  -4.33%  (p=0.000 n=16+16)
EndToEnd/kv-read/vectorize=off/Simple      31.2kB ± 0%    32.2kB ± 0%  +3.34%  (p=0.000 n=18+17)
EndToEnd/kv-read/vectorize=off/Prepared    20.3kB ± 0%    19.5kB ± 0%  -4.08%  (p=0.000 n=16+19)

name                                     old allocs/op  new allocs/op  delta
EndToEnd/kv-read/vectorize=on/Simple          248 ± 1%       249 ± 0%  +0.32%  (p=0.000 n=16+17)
EndToEnd/kv-read/vectorize=on/Prepared        171 ± 1%       157 ± 1%  -8.14%  (p=0.000 n=17+16)
EndToEnd/kv-read/vectorize=off/Simple         242 ± 0%       242 ± 0%  +0.31%  (p=0.000 n=18+17)
EndToEnd/kv-read/vectorize=off/Prepared       164 ± 0%       150 ± 0%  -8.54%  (p=0.000 n=16+16)
```

This prototype also explores a method for utilizing placeholder scans in generic
query plans that don't hit the placeholder fast-path. It does this by adding an
exploration rule to transform simple parameterized lookup joins into placeholder
scans. There are still some wrinkles to iron out, e.g., costing, but I believe
this shows that there is a pragmatic path for supporting compiled factories of
other types of generic query plans. 

Below are some implementation notesfor myself th
Below is a rough map of the steps to productionize this prototype:

#### A. Make placeholder scans a first-class type in exec factories

1. Create a placeholder scan private.
2. Create a placeholder exec params struct.
3. Add placeholder constructor to factories.
4. Move constraint building from execbuilder to factories.
5. Make sure EXPLAIN works.
6. Make sure plan gists work.

#### B. Simple compiled factories for placeholder scans

1. Add simple compilation API to factories.
2. Solve the instrumentation problem.
3. Remove planner from constructPlan.
4. Figure out how to correctly handle codec in a compiled factory.
5. Make NewScanNode a static function.
6. Ensure compilation is only attempted once per generic memo.

#### C. Convert parameterized lookup joins to placeholder scans

1. Add locking to placeholder scan.
2. Add more robust mechanism for unsupported nodes in compiled factories.
3. Add exploration rule to convert parameterized lookup join to placeholder
scan.
4. Add costing for placeholder scans. Maybe this can be really simple for now,
like reuse the cost of the lookup join?

#### D. Support compiled factories for more operators 

1. Project
2. Update
3. Delete
4. Insert?